+++ example/main.dart
@@ -1,8 +1,7 @@
+// ignore_for_file: unused_field
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:google_maps_helper/models/gmh_models.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:google_maps_helper/ui/gmh_map.dart';
-import 'package:google_maps_helper/ui/gmh_search_field.dart';
+import 'package:google_maps_helper/google_maps_helper.dart';
 
 void main() => runApp(const MyApp());
 
@@ -30,55 +29,31 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        useMaterial3: true,
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-      ),
       home: Scaffold(
         appBar: AppBar(title: Text('Google Maps Helper')),
-        body: SafeArea(
-          child: Stack(
-            children: [
-              _map,
-              _searchField,
-            ],
-          ),
-        ),
+        body: SafeArea(child: Stack(children: [_map, _searchField])),
       ),
     );
   }
 
-  /// Google map
+  /// Google map widget
   Widget get _map {
     return GmhMap(
       mapOptions: GmhMapOptions(
         mapType: MapType.normal,
         initialCameraPosition: _kSrc,
-        onTap: (pos) => print('Tapped: $pos'),
+        onTap: (pos) => _getAddress(pos),
         minMaxZoomPreference: MinMaxZoomPreference(15, 18),
         onMapCreated: (controller) => _controller = controller,
         markers: {
-          Marker(
-            draggable: true,
-            position: _kSrc.target,
-            markerId: MarkerId('src'),
-            infoWindow: InfoWindow(title: 'Source'),
-            onTap: () => _controller?.showMarkerInfoWindow(MarkerId('src')),
-          ),
-          Marker(
-            draggable: true,
-            position: _kDest.target,
-            markerId: MarkerId('dest'),
-            infoWindow: InfoWindow(title: 'Destination'),
-            onTap: () => _controller?.showMarkerInfoWindow(MarkerId('dest')),
-          ),
+          Marker(markerId: MarkerId('src'), position: _kSrc.target),
+          Marker(markerId: MarkerId('dest'), position: _kDest.target),
         },
       ),
       polylineOptions: GmhPolylineOptions(
         geodesic: true,
         color: Colors.blue,
+        optimizeWaypoints: true,
         apiKey: '<GOOGLE_DIRECTIONS_API_KEY>',
       ),
     );
@@ -87,9 +62,29 @@ class _MyAppState extends State<MyApp> {
   /// Auto-complete places search field
   Widget get _searchField {
     return GmhSearchField(
-      initialValue: _address,
+      selectedValue: _address,
       onSelected: (data) => _address = data,
       searchParams: GmhSearchParams(apiKey: '<GOOGLE_PLACES_API_KEY>'),
     );
+  }
+
+  /// REVERSE GEOCODING: Get address from geocode
+  Future<void> _getAddress(LatLng pos) async {
+    final data = await GmhService().getAddress(
+      lat: pos.latitude,
+      lng: pos.longitude,
+      apiKey: '<GOOGLE_GEOCODING_API_KEY>',
+    );
+    if (kDebugMode) print(data?.address);
+    if (data != null) _getGeocode(data.address);
+  }
+
+  /// GEOCODING: Get geocode from an address
+  Future<void> _getGeocode(String address) async {
+    final data = await GmhService().getGeocode(
+      address: address,
+      apiKey: '<GOOGLE_GEOCODING_API_KEY>',
+    );
+    if (kDebugMode) print('${data?.lat},${data?.lng}');
   }
 }