+++ lib/google_maps_helper.dart
@@ -1,4 +1,5 @@
 export 'ui/gmh_map.dart';
 export 'models/gmh_models.dart';
 export 'ui/gmh_search_field.dart';
+export 'services/gmh_service.dart';
 export 'package:google_maps_flutter/google_maps_flutter.dart';