+++ lib/services/log_service.dart
@@ -1,3 +1,6 @@
+// import 'dart:developer';
+// import 'package:flutter/foundation.dart';
+
 /// Handle Logs
 class LogService {
   static final _instance = LogService._();