+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- [GEOCODING] Get geocode from an address
+- [REVERSE GEOCODING] Get address from geocode
+- Added polyline param to optimize waypoints
+
 ## 1.0.0
 
 - Auto-complete place search field & polyline drawing